% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_vec.R
\name{cf_vec}
\alias{cf_vec}
\title{Create a cashflow vector}
\usage{
cf_vec(x = double(), freq = 12L)
}
\arguments{
\item{x}{A numeric vector of cashflow values.}

\item{freq}{Periodicity of the cashflows as number of periods per year:
\code{12} (monthly, default), \code{6} (bimonthly), \code{4} (quarterly), \code{3} (trimonthly),
\code{2} (semi-annual), or \code{1} (annual).}
}
\value{
A \code{cf_vec} object.
}
\description{
Create a cashflow vector
}
