% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality_tables}
\alias{mortality_tables}
\alias{qx_iamb}
\alias{scale_g2}
\title{IAM-B mortality rates and Scale G2 improvement factors}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 242 rows and 3 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 242 rows and 3 columns.
}
\source{
\url{https://mort.soa.org}
}
\usage{
qx_iamb

scale_g2
}
\description{
Individual annuity mortality and improvement scales.
}
\details{
\subsection{\code{qx_iamb}}{

IAM-B table mortality rates by age and gender. A tibble with 3 columns:
\itemize{
\item \code{age}: Integer age.
\item \code{qx}: Probability of death within one year.
\item \code{gender}: Factor with levels \code{Female} and \code{Male}.
}
}

\subsection{\code{scale_g2}}{

Scale G2 annual mortality improvement factors by age and gender. A tibble
with 3 columns:
\itemize{
\item \code{age}: Integer age.
\item \code{mi}: Annual mortality improvement rate.
\item \code{gender}: Factor with levels \code{Female} and \code{Male}.
}
}
}
\keyword{datasets}
